Further cleanup, including API changes.
Make numeric types consistent (size_t rather than int or long,
uint8_t for node height).

For all functions that take a callback and a (void *udata),
make the udata be the argument after the callback.

Change _delete API to behave like _get: it should return
a bool and write into an optional *value argument, rather
than returning the value. This makes it possible to distinguish
between a value of NULL and not found.

Optionally compile out skiplist_debug function, so there isn't
a dependency on stdio.

Change release-candidate version to 0.9.0 since the API is
approaching stability.